requirements.txt
# Core Google API
google-api-python-client==2.100.0
google-auth-oauthlib==1.0.0
google-auth==2.23.4

# File extraction
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pdfplumber==0.10.2

# OCR
Pillow==10.0.1
pytesseract==0.3.10

# Utilities
PyYAML==6.0.1

# Dev & testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
